requirements.txt
# Core Dependencies
langgraph>=0.1.0
langchain>=0.1.0
langchain-openai>=0.1.0
openai>=1.0.0

# Kubernetes Integration
kubernetes>=28.0.0
pyyaml>=6.0

# Async and Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
paho-mqtt>=1.6.0

# Monitoring and Tracing
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation>=0.41b0
opentelemetry-exporter-jaeger>=1.21.0
opentelemetry-exporter-otlp>=1.21.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-requests>=0.41b0
opentelemetry-instrumentation-logging>=0.41b0

# Database and State Management
redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Utilities
pydantic>=2.0.0
click>=8.1.0
python-dotenv>=1.0.0
structlog>=23.2.0
tenacity>=8.2.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.6.0

# Optional: GPU and ML Libraries (for advanced scheduling)
torch>=2.0.0
numpy>=1.24.0

# Security
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
